+++ table/src/main/java/tech/ydb/table/rpc/TableRpc.java
@@ -32,6 +32,8 @@ import tech.ydb.proto.table.YdbTable.KeepAliveRequest;
 import tech.ydb.proto.table.YdbTable.KeepAliveResult;
 import tech.ydb.proto.table.YdbTable.PrepareDataQueryRequest;
 import tech.ydb.proto.table.YdbTable.PrepareQueryResult;
+import tech.ydb.proto.table.YdbTable.ReadRowsRequest;
+import tech.ydb.proto.table.YdbTable.ReadRowsResponse;
 import tech.ydb.proto.table.YdbTable.ReadTableRequest;
 import tech.ydb.proto.table.YdbTable.ReadTableResponse;
 import tech.ydb.proto.table.YdbTable.RollbackTransactionRequest;
@@ -156,6 +158,16 @@ public interface TableRpc extends AutoCloseable {
      * @return completable future with result of operation
      */
     CompletableFuture<Result<ExecuteQueryResult>> executeDataQuery(ExecuteDataQueryRequest request,
+            GrpcRequestSettings settings);
+
+    /**
+     * Read rows in the key-value form.
+     * SessionId of previously created session must be provided.
+     * @param request request proto
+     * @param settings rpc call settings
+     * @return completable future with result set
+     */
+    CompletableFuture<Result<ReadRowsResponse>> readRows(ReadRowsRequest request,
             GrpcRequestSettings settings);
 
     /**