+++ table/src/test/java/tech/ydb/table/TableRpcStub.java
@@ -11,6 +11,7 @@ import tech.ydb.core.grpc.GrpcReadStream;
 import tech.ydb.core.grpc.GrpcRequestSettings;
 import tech.ydb.core.impl.call.EmptyStream;
 import tech.ydb.core.utils.Async;
+import tech.ydb.proto.ValueProtos;
 import tech.ydb.proto.table.YdbTable;
 import tech.ydb.proto.table.YdbTable.AlterTableRequest;
 import tech.ydb.proto.table.YdbTable.BeginTransactionRequest;
@@ -34,6 +35,7 @@ import tech.ydb.proto.table.YdbTable.KeepAliveRequest;
 import tech.ydb.proto.table.YdbTable.KeepAliveResult;
 import tech.ydb.proto.table.YdbTable.PrepareDataQueryRequest;
 import tech.ydb.proto.table.YdbTable.PrepareQueryResult;
+import tech.ydb.proto.table.YdbTable.ReadRowsResponse;
 import tech.ydb.proto.table.YdbTable.RollbackTransactionRequest;
 import tech.ydb.table.rpc.TableRpc;
 
@@ -118,6 +120,11 @@ public class TableRpcStub implements TableRpc {
     public CompletableFuture<Result<ExecuteQueryResult>> executeDataQuery(ExecuteDataQueryRequest request,
                                                                                 GrpcRequestSettings settings) {
         return notImplemented("executeDataQuery()");
+    }
+
+    @Override
+    public CompletableFuture<Result<ReadRowsResponse>> readRows(YdbTable.ReadRowsRequest request, GrpcRequestSettings settings) {
+        return notImplemented("readRows()");
     }
 
     @Override